Share one fake window between the analytics and Sentry tests
test-analytics and test-sentry each carried their own makeWindow() with
some twenty identical lines of listener bookkeeping. The shared
test-support/fake-window.mjs gives both consumers the full surface
(removeEventListener, fetch, document.head capture, matchMedia) and the
copy-on-dispatch loop, so a self-removing handler is safe in either.
Also takes webDir/repoRoot/htmlPages from repo-paths instead of
process.cwd().

Model: Claude Fable 5.1 (claude-fable-5-1)
